requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
flask_bcrypt
requests
python-dotenv
email_validator
flask_migrate